Updated command-line parser to accept user information
The parser now accepts a -u or --user option, a -e or --email option, a -m or --major option, and a -n or --number option. This allows us to make sure our samples are unique, and to automate the testing environment by using the -n or --number option to take a pre-determined number of tests, all at once. This information then goes into the file specified by the -o or --output option.