Added phoneme and kv_pair structs with some related functions
The phoneme struct represents phonemes (of course) and is also the key member for kv_pair structs. Related functions are constructors for phonemes and kv_pair structs, as well as a function to hash phoneme structs so that the kv_pair structs containing them may be inserted into a hashmap.