Updated help text to reflect recent changes
The help text (res/help.txt) now tells the user that the program requires superuser permissions, because it reads from device event files in the /dev/input/ directory. The program also explains the usage of new parameters, such as -f, --free-text, -x, --fixed-text, -v, and --device-file; this is requied so new users understand that these are required parameters. Historical parameters have also been updated to include the type of data they expect, for example TEXT when a string is expected, INTEGER when a number is expected, FILE when a file path is expected, and NONE when the parameter takes no value (e.g. the -f and --free-text parameters).