kdtv2 is now just kdt + updated project structure
The build script is also updated to target kdt.c instead of kdtV2.c. The structure of the directories for this project have been modified to put legacy code in the /old directory and experimental code/scripts in the /testing directory. The Makefile has also been removed as we are just using the build bash script located at the base of the project.