Ensured session memory is initialized and fixed error case for set_session_statistic_data
Immediately after the creation of the session structs onto the stack, all of their pointers are set to NULL and all of their length members are set to 0. This memory being formerly unitialized was causing errors in the logic of set_session_statistic_data, which would try to set a variable called statistic_array_length equal to the pointer to the relevant length member (e.g. time_deltas_length when STATISTIC_TIME_DELTAS is passed as the statistic_code parameter). Later when statistic_array_length is deferenced so the value can be set, the progrma would crash, because the memory was unitialized.