Added runner and data-organizer scripts
The runner script will automatically guide the user through the creation of all the data samples that are required of them. It uses a 5 second time incrementor and allows the user to define a starting point and ending point, as well as the number of samples to take for each duration. The runner script produces a unique file for EACH typing exam of any given duration, so it generates many files. This is what the data-organizer script is for; it will move the files into their appropriate child directories within the data directory within the kdt directory (not the data directory at the base of this repo)